test(cli): move the new graph traverser case to Swift Testing
The overlapping-and-chained dynamic xcframework case was added to `GraphTraverserTests`, which is an
XCTest class, so it inherited XCTest. New tests are written with Swift Testing, so it moves to its own
suite. `temporaryPath()` becomes the `.inTemporaryDirectory` trait, and the two assertions become
`#expect`.
